[UPDATE] new features
[UPDATE] new features
Adding new feature : catch header informations that have to be sent when
calling an operation. So if a SOAPHeader is required, it is indicated in
the operation/function php doc (SOAPHeader : required). Moreover, the
names and types of the SOAPHeader(s) are extracted from the WSDL and
added to the operation/function php doc (SOAPHeaderNames and
SOAPHeaderTypes) separated by a comma.
Wsdl location loading has been optimized
Don't add virtual struct (aka restriction and not enumeration) to the
ClassMap array
Adding test case samples
Ensure that meta informations are cleaned and generated correctly either
if it's a string or an array