[UPDATE] new features and php doc improvement
WsdlClass__construct method allows to reset or not the SoapClient if its
not needed.

New methods are generated within the Services classes in order to set
the SoapHeaders after instancitating the service class. For each
required SoapHeader, a method is available and makes easy to set the
SoapHeader.

Improvement in WsdlToPhpModel::getValueWithinItsType() in order to allow
to indicate in wich type the value should be returned

Update Structs __construct() method when WsdlClass is not generated in
order to call its setters