[UPDATE] SoapHeader setters and "from schema" meta
SoapHeader setters are now generated first so the sample file will call
them first as it should be done if you need to set SoapHeaders. They are
also more efficient on namespace definition and on parameters naming.

The meta information "from schema" is introduced in each generated
Struct and Enum class in order to identify the origin of the class. It's
pretty useful when WSDL includes/imports numerous schemas.