[UPDATE] identical operation parameters type naming fix
As the methods, representing the operations, had their parameters named
as their type when the parameter type was an existing known class, it
generated errors as we had multiple parameters with the same name (ex:
login(StructString $_structString,StructString $_structString), which is
conflictual). This was used to generate the sample file easily. This
issue was fixed by naming the parameters with their real name (when
there is multiple parameters) and by reviewing the sample file
generation.