[UPDATE] enhancement and compliance
In order to let the PHP native SoapClient class handle correctly the
object to send based on the request created with the generated classes,
we no longer define the array to send. Indeed, it's better to let the
SoapClient convert the object passed to the generate method mathcing the
operation to call instead of defining an array based on the object. It
actually creates conflicts or defines values that should be sent.
By this update, the option sendArrayAsParameter should be not useful
anymore.
By this update we lose the faculty to define a specific behaviour on the
values sent because we previously called the getter associated to the
value to send. In this case, the generate method PHP doc block has been
simplified.

Update of the sample file to be more explicit on the optional parameter
when instantiating a Service class.

Finally, the sample case for PayPal in samples-generator.php has been
updated to give a functional example based on this update.